% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_blocks.R
\name{as_blocks}
\alias{as_blocks}
\title{Transform a matrix into a list of blocks}
\usage{
as_blocks(M, p)
}
\arguments{
\item{M}{the `matrix` to be transformed in blocks}

\item{p}{an `integer` `vector` with the number of variable in each block}
}
\value{
a named `list` of matrices
}
\description{
Transform a matrix into a list of blocks
}
