% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks_and_variables.R
\name{blocks_and_variables}
\alias{blocks_and_variables}
\title{Creates a tibble with the block and variable names}
\usage{
blocks_and_variables(inputs)
}
\arguments{
\item{inputs}{a list of matrices or data.frames}
}
\value{
a `tibble`
}
\description{
Creates a tibble with the block and variable names
}
