% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_blocks.R
\name{blocks_to_df}
\alias{blocks_to_df}
\title{Convert block lists to a `data.frame`}
\usage{
blocks_to_df(L)
}
\arguments{
\item{L}{a `list` of matrices or data.frames}
}
\value{
a `tibble` (`data.frame`)
}
\description{
Convert block lists to a `data.frame`
}
