% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_Xs_Y.R
\name{compute_Xs_Y}
\alias{compute_Xs_Y}
\title{Computes the simulated data}
\usage{
compute_Xs_Y(
  Bxx = 0,
  Bxy = 0,
  By = 0,
  lX,
  lY,
  p,
  e,
  f,
  snr = 1,
  categorical_Y = FALSE
)
}
\arguments{
\item{Bxx}{(optional) a `nlX x sum(p)` `matrix` specifying the relationship
between the explanatory variables and the X-only latent factors.
Default = no relationship.}

\item{Bxy}{(optional) a `nlY x sum(p)` `matrix` specifying the relationship
between the explanatory variables and the shared latent factors.
Default = no relationship.}

\item{By}{(optional) a `nlY x sum(p)` `matrix` specifying the relationship
between the explanatory variables and the shared latent factors.
Default = no relationship.}

\item{lX}{a `n x nlX` `matrix` with the X-only latent variables.}

\item{lY}{a `n x nlY` `matrix` with the shared latent variables.}

\item{p}{a `vector` of `integer`s with the number of variables in each block.}

\item{e}{a `n x sum(p)` `matrix` with the explanatory variables noise.}

\item{f}{a `n x q` `matrix` with the response variables noise.}

\item{snr}{a strictly positive `double` specifying the signal-to-noise ratio.}

\item{categorical_Y}{}
}
\value{
a `list` with the computed data
}
\description{
Computes the simulated data
}
\keyword{internal}
