% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colors.R
\name{.add_colors}
\alias{.add_colors}
\title{Add colors to a data frame}
\usage{
.add_colors(X, samples_color)
}
\arguments{
\item{X}{a `data.frame`}

\item{sample_color}{a `vector` or a `data.frame` with the colors of the samples}
}
\value{
a `data.frame` object
}
\description{
Add colors to a data frame
}
