% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_blocks.R
\name{.plot_mtb_blocks}
\alias{.plot_mtb_blocks}
\title{Plots the blocks from a list}
\usage{
.plot_mtb_blocks(input_var)
}
\arguments{
\item{input_var}{the output of `blocks_and_variable()`}
}
\value{
a `ggplot2` object
}
\description{
Plots the blocks from a list
}
\keyword{internal}
