% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_data.R
\name{get_block_data}
\alias{get_block_data}
\title{Creates a tibble with the block and variable names}
\usage{
get_block_data(inputs)
}
\arguments{
\item{inputs}{a list of matrices or data.frames}
}
\value{
a `tibble`
}
\description{
Creates a tibble with the block and variable names
}
