% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtb_bipc.R
\name{get_mtb_bipc}
\alias{get_mtb_bipc}
\title{Retrieve and format the observed (and bootstrapped) cumulative Block Importance in Projections (BIPC)}
\usage{
get_mtb_bipc(res, boot = NULL, CI = 0.95)
}
\arguments{
\item{res}{the output of `ade4::mbpls` (or `ade4::mbpcaiv`)}

\item{boot}{(optional) the output of `ade4::randboot`}

\item{CI}{The confidence interval that should be displayed. Default is 0.95.}

\item{wrap_block_names}{(optional) `logical` specifying if block names (x-axis) should be wrapped. Default is `TRUE`.}
}
\value{
a `tibble` object
}
\description{
Retrieve and format the observed (and bootstrapped) cumulative Block Importance in Projections (BIPC)
}
