% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_data.R
\name{get_mtb_blocks}
\alias{get_mtb_blocks}
\title{Extract block data into long format}
\usage{
get_mtb_blocks(blocks)
}
\arguments{
\item{blocks}{}
}
\value{
a `tibble`
}
\description{
Extract block data into long format
}
\keyword{internal}
