% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtb_coef.R
\name{get_mtb_coef}
\alias{get_mtb_coef}
\title{Retrieve the coefficients of associations between the explanatory and response variables}
\usage{
get_mtb_coef(res, boot = NULL, Y_var = NULL, CI = 0.95, as_matrix = FALSE)
}
\arguments{
\item{res}{the output of `ade4::mbpls` (or `ade4::mbpcaiv`) (the input of `ade4::randboot`)}

\item{boot}{(optional) the output of `ade4::randboot` to display the confidence intervals}

\item{Y_var}{(optional) a `integer` or a `character` specifying the indices
or names of selected response variables if one desires to display the
coefficients for a subset of variables.}

\item{CI}{(optional) The width of the confidence interval that should be displayed. Default is 0.95.}

\item{as_matrix}{(optional) a `logical` specifying if the coefficients have to be returned in matrix form (n_input x n_output)}
}
\value{
a `tibble`
}
\description{
Retrieve the coefficients of associations between the explanatory and response variables
}
