% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtb_eig.R
\name{get_mtb_eig}
\alias{get_mtb_eig}
\title{Extract Eigenvalues from a `ade4::mbpls` or `ade4::mbpcaiv` result}
\usage{
get_mtb_eig(res)
}
\arguments{
\item{res}{the output of `ade4::mbpls` or `ade4::mbpcaiv`}
}
\value{
a `tibble`
}
\description{
Extract Eigenvalues from a `ade4::mbpls` or `ade4::mbpcaiv` result
}
