% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_Xloadings.R
\name{get_mtb_loadings}
\alias{get_mtb_loadings}
\title{Retrieve the variables loading as tibble}
\usage{
get_mtb_loadings(res, cut_at_nf = TRUE)
}
\arguments{
\item{res}{the output of `ade4::mbpls`, `ade4::mbpcaiv`, or `packMBPLSDA::mbplsda`}

\item{cut_at_nf}{a `logical` indicating whether to only include the first `res$nf` latent dimensions (`TRUE`) or all available latent dimensions (`FALSE`). Default is `TRUE`.}
}
\value{
a `tibble` object
}
\description{
Retrieve the variables loading as tibble
}
