% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtb_vipc.R
\name{get_mtb_vipc}
\alias{get_mtb_vipc}
\title{Retrieve and format the observed (and bootstrapped) cumulative Variable Importance in Projection (VIPC)}
\usage{
get_mtb_vipc(res, boot = NULL, CI = 0.95)
}
\arguments{
\item{res}{the output of `ade4::mbpls` (or `ade4::mbpcaiv`)}

\item{boot}{(optional) the output of `ade4::randboot`}

\item{CI}{(optional) The width of the confidence interval that should be displayed. Default is 0.95.}
}
\value{
a `tibble` object
}
\description{
Retrieve and format the observed (and bootstrapped) cumulative Variable Importance in Projection (VIPC)
}
