% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtbda_coef.R
\name{get_mtbda_coef}
\alias{get_mtbda_coef}
\title{Retrieve the coefficients of associations between the explanatory and response variables for a MBPLSDA object}
\usage{
get_mtbda_coef(res, boot = NULL, Y_var = NULL, CI = 0.95, as_matrix = FALSE)
}
\arguments{
\item{res}{the output of `packMBPLSDA::mbplsda` (the input of `packMBPLSDA::boot_mbplsda`)}

\item{boot}{(optional) the output of `packMBPLSDA::boot_mbplsda` to display the confidence intervals}

\item{Y_var}{(optional) a `integer` or a `character` specifying the indices
or names of selected response variables if one desires to display the
coefficients for a subset of variables.}

\item{as_matrix}{(optional) a `logical` specifying if the coefficients have to be returned in matrix form (n_input x n_output)}
}
\value{
a `tibble`
}
\description{
Retrieve the coefficients of associations between the explanatory and response variables for a MBPLSDA object
}
