% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refs.R
\name{get_ref_bipc}
\alias{get_ref_bipc}
\title{Compute reference cummulative block importance importance}
\usage{
get_ref_bipc(res)
}
\arguments{
\item{res}{the output of `ade4::mbpls` (or `ade4::mbpcaiv` or `packMBPLSDA::mbplsda`)}
}
\value{
a `tibble` object
}
\description{
Compute reference cummulative block importance importance
}
