% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_axis.R
\name{mirror_axis}
\alias{mirror_axis}
\title{Mirror selected axes from an `ade4::mbpls` model}
\usage{
mirror_axis(res, axis = 1)
}
\arguments{
\item{res}{the output of `ade4::mbpls`}

\item{axis}{the axis that needs to be mirrored (*-1)}
}
\value{
a list similar to the output of `ade4::mbpls`
}
\description{
Useful for comparing several models
}
