% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtb_predict.R
\name{mtb_predict}
\alias{mtb_predict}
\title{Predict Y from an `ade4::mbpls` model}
\usage{
mtb_predict(res, newdata = NULL)
}
\arguments{
\item{res}{the output of an `ade4::mbpls` call}

\item{newdata}{(optional)}
}
\value{
a matrix
}
\description{
Predict Y from an `ade4::mbpls` model
}
