% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_blocks.R
\name{plot_mtb_blocks}
\alias{plot_mtb_blocks}
\title{Displays the blocks structure}
\usage{
plot_mtb_blocks(Xs, Y)
}
\arguments{
\item{Xs}{a `list` with the explanatory blocks}

\item{Y}{a single element `list` with the response block}
}
\value{
a `ggplot2` object
}
\description{
Displays the blocks structure
}
