% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_cov.R
\name{plot_mtb_cov}
\alias{plot_mtb_cov}
\title{Plots block covariances}
\usage{
plot_mtb_cov(res, xaxis = 1, yaxis = 2, scale_axes = "eig")
}
\arguments{
\item{res}{the output of `ade4::mbpls` or `ade4::mbpcaiv`}

\item{xaxis}{an `integer` specifying which latent variable should be plotted on the x-axis. Default is 1.}

\item{yaxis}{an `integer` specifying which latent variable should be plotted on the y-axis. Default is 1.}

\item{scale_axes}{how the aspect ratio between the x-axis and y-axis should be computed (default is "fixed", other option is "eig"). XXX explain better XXX}
}
\value{
a `ggplot2` object
}
\description{
Plots block covariances
}
