% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_data.R
\name{plot_mtb_data}
\alias{plot_mtb_data}
\title{Plot the explanatory and response variables}
\usage{
plot_mtb_data(Xs, Y)
}
\arguments{
\item{Xs}{a `list` with the explanatory blocks}

\item{Y}{a single element `list` with the response block}
}
\value{
a `patchwork` of `ggplot2` object
}
\description{
Plot the explanatory and response variables
}
