% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_lY.R
\name{plot_mtb_lY}
\alias{plot_mtb_lY}
\title{Plot the projections of the response variables onto the latent space}
\usage{
plot_mtb_lY(
  res,
  xaxis = 1,
  yaxis = 2,
  scale_axes = "eig",
  add_loadings = TRUE,
  samples_color = NULL
)
}
\arguments{
\item{res}{the output of `ade4::mbpls` or `ade4::mbpcaiv`}

\item{xaxis}{an `integer` specifying which latent variable should be plotted on the x-axis. Default is 1.}

\item{yaxis}{an `integer` specifying which latent variable should be plotted on the y-axis. Default is 1.}

\item{scale_axes}{how the aspect ratio between the x-axis and y-axis should be computed (default is "fixed", other option is "eig"). XXX explain better XXX}

\item{add_loadings}{(optional) a `logical` specifying whether the response loadings should be added to create a biplot.}

\item{samples_color}{(optional) a `vector` or a `matrix` specifying how samples should be colored. XXX give more explanations XXX}
}
\value{
a `ggplot2` object
}
\description{
Plot the projections of the response variables onto the latent space
}
