% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_vipc.R
\name{plot_mtb_vipc}
\alias{plot_mtb_vipc}
\title{Plots the bootstraped and observed Block Importance}
\usage{
plot_mtb_vipc(boot, input_var)
}
\arguments{
\item{boot}{the output of `ade4::randboot`}

\item{input_var}{a `data.frame` xxxx}
}
\value{
a `ggplot2` object
}
\description{
Plots the bootstraped and observed Block Importance
}
