% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_varX_varY.R
\name{plot_varX_varY}
\alias{plot_varX_varY}
\title{Scatter plot of two variables against each other}
\usage{
plot_varX_varY(Xs = NULL, Y = NULL, res = NULL, varX, varY, color_by = NULL)
}
\arguments{
\item{Xs}{(optional) the multiblock input (list of `matrices`/`data.frames`)}

\item{Y}{(optional) the output `matrix`/`data.frame`}

\item{res}{(optional - either `res` or `Xs` and `Y`) the result of `ade4::mbpls` or `ade4::mbpcaiv`}

\item{varX}{a character specifying the x-axis variable (can be an explanatory or response variable)}

\item{varY}{a character specifying the y-axis variable (can be an explanatory or response variable)}

\item{color_by}{(optional) a character or a vector of the same length of the data
specifying how the samples should be colored}
}
\value{
a `ggplot2` object
}
\description{
Scatter plot of two variables against each other
}
