% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_nf.R
\name{retrieve_nf}
\alias{retrieve_nf}
\title{Retrieve the number of latent factors used for the bootstrap CI estimations}
\usage{
retrieve_nf(res, boot)
}
\arguments{
\item{res}{the output of `ade4::mbpls` (or `ade4::mbpcaiv`) (the input of `ade4::randboot`)}

\item{boot}{(optional) the output of `ade4::randboot` to display the confidence intervals}
}
\value{
an `integer`
}
\description{
Retrieve the number of latent factors used for the bootstrap CI estimations
}
\keyword{internal}
