% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_simulated_latent_structure.R
\name{visualize_simulated_latent_structure}
\alias{visualize_simulated_latent_structure}
\title{Plot the simulated latent structures}
\usage{
visualize_simulated_latent_structure(A)
}
\arguments{
\item{A}{the output of `simulate_Xs_Y`}
}
\value{
a `ggplot2` object
}
\description{
Plot the simulated latent structures
}
