% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtb_predict.R
\name{mtb_predict}
\alias{mtb_predict}
\title{Predict Y from an `ade4::mbpls` model}
\usage{
mtb_predict(res, original_Y = NULL, newdata = NULL, original_X = NULL)
}
\arguments{
\item{res}{the output of an `ade4::mbpls` or `packMBPLSDA::mbplsda` call}

\item{original_Y}{(optional) the `matrix` or `data.frame` that was provided to the `res` model.
If not provided, the predictions will be centered and scaled.}

\item{newdata}{(optional) an `ade4::ktab` object - similar to what would be
provided as input of `ade4::mbpls`.
If `newdata` is not provided, the predictions for the original data are provided.}

\item{original_X}{(mandatory if newdata is not null) the original `ade4::ktab`
object on whicht the model (`res`) was fitted}
}
\value{
a `matrix` with the predicted output.
}
\description{
Predict Y from an `ade4::mbpls` model
}
