% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_bipc.R
\name{plot_mtb_bipc}
\alias{plot_mtb_bipc}
\title{Plots the bootstraped and observed Block Importance}
\usage{
plot_mtb_bipc(boot, input_var, show_dist = FALSE, CI = 0.95)
}
\arguments{
\item{boot}{the output of `ade4::randboot`}

\item{input_var}{a `data.frame` xxxx}

\item{show_dist}{(optional) a `logical` specifying if the bootstrap distribution should be shown with a boxplot.
Default is FALSE}

\item{CI}{The confidence interval that should be displayed. Default is 0.95.}
}
\value{
a `ggplot2` object
}
\description{
Plots the bootstraped and observed Block Importance
}
