% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_bipc.R
\name{plot_mtb_bipc}
\alias{plot_mtb_bipc}
\title{Plots the observed (and bootstrapped) cumulative Block Importance in Projections (BIPC)}
\usage{
plot_mtb_bipc(
  res,
  boot = NULL,
  show_dist = FALSE,
  CI = 0.95,
  wrap_block_names = TRUE
)
}
\arguments{
\item{res}{the output of `ade4::mbpls` (or `ade4::mbpcaiv`)}

\item{boot}{(optional) the output of `ade4::randboot`}

\item{show_dist}{(optional) a `logical` specifying if the bootstrap distribution should be shown with a boxplot.
Default is FALSE}

\item{CI}{The confidence interval that should be displayed. Default is 0.95.}

\item{wrap_block_names}{(optional) `logical` specifying if block names (x-axis) should be wrapped. Default is `TRUE`.}
}
\value{
a `ggplot2` object
}
\description{
Plots the observed (and bootstrapped) cumulative Block Importance in Projections (BIPC)
}
