% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_bipc.R
\name{plot_mtb_bipc}
\alias{plot_mtb_bipc}
\title{Plots the observed (and bootstrapped) cumulative Block Importance in Projections (BIPC)}
\usage{
plot_mtb_bipc(
  res,
  boot = NULL,
  show_dist = FALSE,
  CI = 0.95,
  show_ref = TRUE,
  show_ref_ratio = FALSE,
  wrap_block_names = NULL
)
}
\arguments{
\item{res}{the output of `ade4::mbpls` (or `ade4::mbpcaiv`)}

\item{boot}{(optional) the output of `ade4::randboot`}

\item{show_dist}{(optional) a `logical` specifying if the bootstrap distribution should be shown with a boxplot.
Default is FALSE}

\item{CI}{The confidence interval that should be displayed. Default is 0.95.}

\item{show_ref}{(optional) a `logical` specifying if the reference value under the assumption of equal variable importance should be shown. Default is `TRUE`.}

\item{wrap_block_names}{(optional) `integer` specifying if the block names should be wrapped to the specified character length. Default is `NULL`, indicating no wrapping.}
}
\value{
a `ggplot2` object
}
\description{
Plots the observed (and bootstrapped) cumulative Block Importance in Projections (BIPC)
}
