% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_blocks.R
\name{plot_mtb_blocks}
\alias{plot_mtb_blocks}
\title{Displays the blocks structure}
\usage{
plot_mtb_blocks(
  Xs,
  Y,
  response_first = FALSE,
  min_text_size = 3,
  max_text_size = 4,
  title_sep = "\\n",
  add_n = TRUE
)
}
\arguments{
\item{Xs}{a `list` with the explanatory blocks}

\item{Y}{a single element `list` with the response block}

\item{response_first}{(optional) a `logical` specifying if the response block
should be on the left (`response_first = TRUE`) or on the right
(`response_first = FALSE`, default) of the explanatory variables.}

\item{min_text_size}{(optional) a positive `double` specifying the minimum
for the block variable text size.}

\item{max_text_size}{(optional) a positive `double` specifying the maximum
for the block variable text size.}

\item{title_sep}{(optional) a `character` specifying the separating character
for the title (default is `"\n"`).}

\item{add_n}{(optional) a `logical` specifying if the number of samples should
be specified on the y-axis (default is `TRUE`)}
}
\value{
a `ggplot2` object
}
\description{
Displays the blocks structure
}
