% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_coef.R
\name{plot_mtb_coef}
\alias{plot_mtb_coef}
\title{Plots the coefficients of associations between the explanatory and response variables}
\usage{
plot_mtb_coef(
  boot,
  Y_var = NULL,
  input_var,
  CI = 0.95,
  format = "long",
  max_coef_value = NULL
)
}
\arguments{
\item{boot}{the output of `ade4::randboot`}

\item{Y_var}{(optional) a `character` specifying the names of selected response variables
if one desires to display the coefficients for a subset of variables.}

\item{input_var}{a `data.frame` xxxx}

\item{CI}{The width of the confidence interval that should be displayed. Default is 0.95.}

\item{format}{"long" or "wide". a `character` specifying if the coefficients should be displayed horizontally or vertically. (default = "long")}

\item{max_coef_value}{an optional upper limit on the coefficient value displayed (to crop the axes).}
}
\value{
a `ggplot2` object
}
\description{
Plots the coefficients of associations between the explanatory and response variables
}
