% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_eig.R
\name{plot_mtb_eig}
\alias{plot_mtb_eig}
\title{Plots the eigenvalues and the cummulative projected inertia}
\usage{
plot_mtb_eig(res, xaxis = NULL, yaxis = NULL, annot = TRUE)
}
\arguments{
\item{res}{the output of `ade4::mbpls` or `ade4::mbpcaiv`}

\item{xaxis}{(optional) an `integer` specifying which latent variable to highlight}

\item{yaxis}{(optional) an `integer` specifying which latent variable should be plotted on the y-axis. Default is 1.}
}
\value{
a `ggplot2` object
}
\description{
Plots the eigenvalues and the cummulative projected inertia
}
