% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_vipc.R
\name{plot_mtb_vipc}
\alias{plot_mtb_vipc}
\title{Plots the observed (and bootstrapped) cumulative Variable Importance in Projection (VIPC)}
\usage{
plot_mtb_vipc(res, boot = NULL, CI = 0.95, show_ref = TRUE)
}
\arguments{
\item{res}{the output of `ade4::mbpls` (or `ade4::mbpcaiv`)}

\item{boot}{(optional) the output of `ade4::randboot`}

\item{CI}{(optional) The width of the confidence interval that should be displayed. Default is 0.95.}

\item{show_ref}{(optional) a `logical` specifying if the reference value under the assumption of equal variable importance within blocks should be shown. Default is `TRUE`.}
}
\value{
a `ggplot2` object
}
\description{
Plots the observed (and bootstrapped) cumulative Variable Importance in Projection (VIPC)
}
