% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtb_Xloadings.R
\name{plot_mtb_Xloadings}
\alias{plot_mtb_Xloadings}
\title{Plot the projections of the input variable onto the latent space}
\usage{
plot_mtb_Xloadings(res, input_var, xaxis = 1, yaxis = 2, scale_axes = "eig")
}
\arguments{
\item{res}{the output of `ade4::mbpls` or `ade4::mbpcaiv`}

\item{input_var}{a `data.frame` xxxx}

\item{xaxis}{an `integer` specifying which latent variable should be plotted on the x-axis. Default is 1.}

\item{yaxis}{an `integer` specifying which latent variable should be plotted on the y-axis. Default is 1.}

\item{scale_axes}{how the aspect ratio between the x-axis and y-axis should be computed (default is "fixed", other option is "eig"). XXX explain better XXX}
}
\value{
a `ggplot2` object
}
\description{
Plot the projections of the input variable onto the latent space
}
